cleanup the codespace and update readme
