+++ .gitignore
@@ -1,4 +1,5 @@
 __pycache__
-venv/
+*venv/
 .pass
 example_cookie.txt
+session.json
