--- QUICK_START.md
@@ -1,28 +1,0 @@
-# 🚀 快速开始指南
-
-## 🌐 快速开始
-
-```bash
-# 1. 并发下载所有资源
-python learn_async.py --multi
-
-# 2. 下载特定课程
-python learn_async.py --course "课程名称"
-
-# 3. 指定并发数
-python learn_async.py --multi --processes 4
-```
-
-## 📋 使用流程
-
-1. **运行命令** - 程序会自动打开 Chrome 浏览器
-2. **手动登录** - 在浏览器中完成登录（支持双因素认证）
-3. **自动继续** - 登录成功后程序自动提取会话并继续
-4. **开始下载** - 程序会显示课程列表并开始下载
-
-## ⚠️ 注意事项
-
-- **首次使用需要安装 Chrome 浏览器**
-- 确保网络连接正常
-- **登录时不要关闭浏览器窗口**
-- 支持所有原有功能（多进程、课程筛选等）
