+++ requirements.txt
@@ -2,3 +2,4 @@ bs4
 tqdm
 requests
 gmssl
+selenium
